[Refactoring] split module
